chore: pluginified the schema
